+++ web/apps/photos/src/utils/publicCollectionGallery/index.ts
@@ -1,17 +1,21 @@
+import { TimeStampListItem } from "components/PhotoList";
 import { createContext } from "react";
-import { PublicCollectionGalleryContextType } from "types/publicCollection";
 
-const defaultPublicCollectionGalleryContext: PublicCollectionGalleryContextType =
-    {
+export interface PublicCollectionGalleryContextType {
+    token: string;
+    passwordToken: string;
+    referralCode: string | null;
+    accessedThroughSharedURL: boolean;
+    photoListHeader: TimeStampListItem;
+    photoListFooter: TimeStampListItem;
+}
+
+export const PublicCollectionGalleryContext =
+    createContext<PublicCollectionGalleryContextType>({
         token: null,
         passwordToken: null,
         referralCode: null,
         accessedThroughSharedURL: false,
         photoListHeader: null,
         photoListFooter: null,
-    };
-
-export const PublicCollectionGalleryContext =
-    createContext<PublicCollectionGalleryContextType>(
-        defaultPublicCollectionGalleryContext,
-    );
+    });
